Pull from master

+++ cruise-control-core/src/main/java/com/linkedin/cruisecontrol/detector/metricanomaly/MetricAnomaly.java
@@ -28,5 +28,5 @@ public interface MetricAnomaly<E extends Entity> extends Anomaly {
   /**
    * Get the metric Id caused the metric anomaly.
    */
-  Integer metricId();
+  Short metricId();
 }
